Refactor dns plugin
Our use of moby/moby for generating the resolvconf does not work after
v26 upgrade because of their use of internal package.

We only use the upstream resolvconf.Build method, which is
self-contained so this commit copies the upstream method to the dns
plugin and removes the use of github.com/moby/moby module.

Signed-off-by: Fredrik Lönnegren <fredrik@frelon.se>